add mock-snos-from-pie for dev purposes
For a quick dev to speed up testing,
this implementation allows to mock snos proof from the PIE content.
This is not intended to be a final solution, and having a
MockSnosProver will be a better long-term solution.
