+++ saya/core/src/prover/atlantic/snos.rs
@@ -3,6 +3,7 @@ use std::{io::Write, time::Duration};
 use anyhow::Result;
 use cairo_vm::vm::runners::cairo_pie::CairoPie;
 use log::{debug, info, trace};
+use starknet::core::types::Felt;
 use tokio::sync::mpsc::{Receiver, Sender};
 use zip::{write::FileOptions, ZipWriter};
 
@@ -16,6 +17,7 @@ use crate::{
         Prover, ProverBuilder, SnosProof,
     },
     service::{Daemon, FinishHandle, ShutdownHandle},
+    utils::{compute_program_hash_from_pie, extract_pie_output, stark_proof_mock},
 };
 
 const PROOF_STATUS_POLL_INTERVAL: Duration = Duration::from_secs(10);
@@ -28,6 +30,8 @@ pub struct AtlanticSnosProver<P> {
     statement_channel: Receiver<NewBlock>,
     proof_channel: Sender<SnosProof<P>>,
     finish_handle: FinishHandle,
+    /// Whether to extract the output and compute the program hash from the PIE or use the one from the SHARP bootloader returned by the prover service.
+    mock_snos_from_pie: bool,
 }
 
 #[derive(Debug)]
@@ -35,6 +39,7 @@ pub struct AtlanticSnosProverBuilder<P> {
     api_key: String,
     statement_channel: Option<Receiver<NewBlock>>,
     proof_channel: Option<Sender<SnosProof<P>>>,
+    mock_snos_from_pie: bool,
 }
 
 impl<P> AtlanticSnosProver<P>
@@ -53,6 +58,26 @@ where
             // This should be fine for now as block ingestors wouldn't drop senders. This might
             // change in the future.
             let new_block = new_block.unwrap();
+
+            // TODO: move this to a separate MockSnosProver.
+            if self.mock_snos_from_pie {
+                let output = bootloader_snos_output(&new_block.pie);
+                let mock_proof = stark_proof_mock(&output);
+
+                info!(
+                    "Mock proof generated from PIE for block #{}",
+                    new_block.number
+                );
+
+                let new_proof = SnosProof {
+                    block_number: new_block.number,
+                    proof: P::parse(serde_json::to_string(&mock_proof).unwrap()).unwrap(),
+                };
+
+                let _ = self.proof_channel.send(new_proof).await;
+
+                continue;
+            }
 
             trace!("Compressing PIE for block #{}", new_block.number);
 
@@ -129,11 +154,12 @@ where
 }
 
 impl<P> AtlanticSnosProverBuilder<P> {
-    pub fn new(api_key: String) -> Self {
+    pub fn new(api_key: String, mock_snos_from_pie: bool) -> Self {
         Self {
             api_key,
             statement_channel: None,
             proof_channel: None,
+            mock_snos_from_pie,
         }
     }
 }
@@ -154,6 +180,7 @@ where
                 .proof_channel
                 .ok_or_else(|| anyhow::anyhow!("`proof_channel` not set"))?,
             finish_handle: FinishHandle::new(),
+            mock_snos_from_pie: self.mock_snos_from_pie,
         })
     }
 
@@ -212,4 +239,24 @@ fn compress_pie(pie: &CairoPie) -> std::result::Result<Vec<u8>, std::io::Error> 
     zip_writer.finish()?;
 
     Ok(bytes.into_inner())
+}
+
+/// Mocks a bootloaded execution of SNOS.
+fn bootloader_snos_output(pie: &CairoPie) -> Vec<Felt> {
+    let snos_program_hash =
+        compute_program_hash_from_pie(pie).expect("Failed to compute program hash from PIE");
+    debug!("SNOS program hash from PIE: {:x}", snos_program_hash);
+
+    let snos_output = extract_pie_output(pie);
+
+    let mut bootloader_output = vec![
+        // Bootloader config (not checked by piltover, set to 0)
+        Felt::ZERO,
+        // bootloader output len (not checked by piltover, set to 0)
+        Felt::ZERO,
+        snos_program_hash,
+    ];
+
+    bootloader_output.extend(snos_output);
+    bootloader_output
 }
